Refactor plugin internals so google-services plugin can reuse via dependency.
